Skip blast tests if not on users; sample parser tests actually test database adding.

+++ src/geovision/text_to_db/tests/blast_parser_tests.py
@@ -1,19 +1,21 @@
 from decimal import *
-import unittest
+from django.utils import unittest
 
-from geovision.settings import TEST_FILE_PATH
+from geovision.settings import TEST_FILE_PATH, RUNNING_ON_USERS
 from geovision.text_to_db.blast_parser import create_blast
 from geovision.viz.models import Blast
 from geovision.viz.models import DbEntry
 from geovision.viz.models import Read
 from nose.tools import raises
+#from nose.plugins.skip.Skip import SkipTest
 
+@unittest.skipUnless(RUNNING_ON_USERS, "Blast parser tests require a PostgreSQL database")
 class BlastParserTests(unittest.TestCase):
-	
+
 	@classmethod
 	def setUpClass(cls):
-
-		file = open(TEST_FILE_PATH + "test_blast.txt")
+		if not RUNNING_ON_USERS:
+			return
 		#in case objects left from other tests..
 		DbEntry.objects.all().delete()
 		Read.objects.all().delete()
@@ -30,6 +32,8 @@ class BlastParserTests(unittest.TestCase):
 		create_blast(cls.dbe[0].source_file, cls.read.sample, TEST_FILE_PATH + "test_blast.txt")
 
 	def test_parse_blast(self):
+		
+
 		results = Blast.objects.all()
 		for i in range(len(results)):
 			self.assertEqual(results[i].read, self.read)
